+++ src/main/java/com/aetherteam/aether/item/combat/abilities/weapon/WeaponAbilityHooks.java
@@ -1,4 +1,4 @@
-package com.aetherteam.aether.event.hooks;
+package com.aetherteam.aether.item.combat.abilities.weapon;
 
 import com.aetherteam.aether.AetherConfig;
 import com.aetherteam.aether.AetherTags;