+++ src/main/java/com/aetherteam/aether/item/tools/holystone/HolystonePickaxeItem.java
@@ -6,7 +6,7 @@ import net.minecraft.world.item.Item;
 import net.minecraft.world.item.PickaxeItem;
 
 /**
- * Ambrosium dropping behavior is called by {@link com.aetherteam.aether.event.hooks.ToolAbilityHooks#handleHolystoneToolAbility}.
+ * Ambrosium dropping behavior is called by {@link com.aetherteam.aether.item.tools.abilities.ToolAbilityHooks#handleHolystoneToolAbility}.
  */
 public class HolystonePickaxeItem extends PickaxeItem implements HolystoneTool {
     public HolystonePickaxeItem() {